# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Blaaize Wallet API'
  description: 'This is the API documentation for Blaaize Wallet.'
  version: 1.0.0
paths:
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: 'View all users'
      description: 'Retrieve a list of all users, filterable by name, email, and creation date.'
      operationId: 12d8068fb9094b67b1d7144ae2ee938d
      parameters:
        -
          name: firstname
          in: query
          description: 'Filter by firstname'
          required: false
          schema:
            type: string
        -
          name: lastname
          in: query
          description: 'Filter by lastname'
          required: false
          schema:
            type: string
        -
          name: email
          in: query
          description: 'Filter by email'
          required: false
          schema:
            type: string
        -
          name: created_at
          in: query
          description: 'Filter by creation date'
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: 'List of users retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        firstname:
                          type: string
                          example: John
                        lastname:
                          type: string
                          example: Doe
                        email:
                          type: string
                          format: email
                          example: johndoe@example.com
                      type: object
                  message:
                    type: string
                    example: 'Users list retrieved successfully'
                type: object
        401:
          description: Unauthenticated
        500:
          description: 'Internal Server Error'
  '/api/admin/user/{id}':
    get:
      tags:
        - Admin
      summary: 'View single user profile'
      description: "Retrieve a user's profile including their wallets and recent transactions."
      operationId: 977c243bb654efcc9c0383c0d6ed53cb
      parameters:
        -
          name: id
          in: path
          description: 'User ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'User profile retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      user:
                        properties:
                          id:
                            type: integer
                            example: 1
                          firstname:
                            type: string
                            example: John
                          lastname:
                            type: string
                            example: Doe
                          email:
                            type: string
                            format: email
                            example: johndoe@example.com
                        type: object
                      wallets:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              example: 1
                            currency:
                              type: string
                              example: USD
                            balance:
                              type: number
                              example: 100.5
                          type: object
                      transactions:
                        type: array
                        items:
                          properties:
                            transaction_id:
                              type: integer
                              example: 123
                            amount:
                              type: number
                              example: 50.75
                            status:
                              type: string
                              example: Completed
                            date:
                              type: string
                              format: date-time
                              example: '2025-03-15T14:00:00Z'
                          type: object
                    type: object
                  message:
                    type: string
                    example: 'User profile retrieved successfully'
                type: object
        404:
          description: 'User not found'
        500:
          description: 'Internal Server Error'
      security:
        -
          sanctum: []
  /api/admin/transactions:
    get:
      tags:
        - Admin
      summary: 'Get all transactions'
      operationId: 70f65b295a9ffc832682153892d19184
      parameters:
        -
          name: user_id
          in: query
          description: 'Filter by user ID'
          required: false
          schema:
            type: integer
        -
          name: currency
          in: query
          description: 'Filter by currency'
          required: false
          schema:
            type: string
        -
          name: date_from
          in: query
          description: 'Filter from date (YYYY-MM-DD)'
          required: false
          schema:
            type: string
            format: date
        -
          name: date_to
          in: query
          description: 'Filter to date (YYYY-MM-DD)'
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: 'Successful response'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        transaction_id:
                          type: integer
                          example: 123
                        user_id:
                          type: integer
                          example: 1
                        amount:
                          type: number
                          example: 150.75
                        currency:
                          type: string
                          example: USD
                        status:
                          type: string
                          example: Completed
                        date:
                          type: string
                          format: date-time
                          example: '2025-03-15T14:00:00Z'
                      type: object
                  message:
                    type: string
                    example: 'Transactions retrieved successfully'
                type: object
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
  '/api/admin/transaction/{id}':
    get:
      tags:
        - Admin
      summary: 'Get transaction by ID'
      operationId: 76e1f2b0465811ea1e58377adab62843
      parameters:
        -
          name: id
          in: path
          description: 'Transaction ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful response'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      transaction_id:
                        type: integer
                        example: 123
                      user_id:
                        type: integer
                        example: 1
                      amount:
                        type: number
                        example: 150.75
                      currency:
                        type: string
                        example: USD
                      status:
                        type: string
                        example: Completed
                      date:
                        type: string
                        format: date-time
                        example: '2025-03-15T14:00:00Z'
                    type: object
                  message:
                    type: string
                    example: 'Transaction retrieved successfully'
                type: object
        404:
          description: 'Transaction not found'
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
  /api/admin/exchange-rates:
    get:
      tags:
        - Admin
      summary: 'Get all exchange rates'
      operationId: 8f3c24adff6922bc763a57bb04b6df37
      responses:
        200:
          description: 'Successful response'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        currency:
                          type: string
                          example: USD
                        rate:
                          type: number
                          example: 1.2345
                        updated_at:
                          type: string
                          format: date-time
                          example: '2025-03-15T14:00:00Z'
                      type: object
                  message:
                    type: string
                    example: 'Exchange rates retrieved successfully'
                type: object
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
  /api/admin/exchange-rate:
    post:
      tags:
        - Admin
      summary: 'Create a new exchange rate'
      operationId: 9e827ebc7a4cea318051baf7b1523f7a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - from_currency
                - to_currency
                - rate
              properties:
                from_currency:
                  type: string
                  example: USD
                to_currency:
                  type: string
                  example: EUR
                rate:
                  type: number
                  format: float
                  example: 1.02
              type: object
      responses:
        201:
          description: 'Exchange rate created successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      from_currency:
                        type: string
                        example: USD
                      to_currency:
                        type: string
                        example: EUR
                      rate:
                        type: number
                        format: float
                        example: 1.02
                      created_at:
                        type: string
                        format: date-time
                        example: '2025-03-15T14:00:00Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2025-03-15T14:00:00Z'
                    type: object
                  message:
                    type: string
                    example: 'Exchange rate created successfully'
                type: object
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
  '/api/admin/exchange-rate/{id}':
    put:
      tags:
        - Admin
      summary: 'Update an exchange rate'
      operationId: 23445238668e08c3891c01c270953ce2
      parameters:
        -
          name: id
          in: path
          description: 'Exchange rate ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - rate
              properties:
                rate:
                  type: number
                  format: float
                  example: 1.02
              type: object
      responses:
        200:
          description: 'Exchange rate updated successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      from_currency:
                        type: string
                        example: USD
                      to_currency:
                        type: string
                        example: EUR
                      rate:
                        type: number
                        format: float
                        example: 1.02
                      updated_at:
                        type: string
                        format: date-time
                        example: '2025-03-15T14:00:00Z'
                    type: object
                  message:
                    type: string
                    example: 'Exchange rate updated successfully'
                type: object
        404:
          description: 'Exchange rate not found'
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Admin
      summary: 'Delete an exchange rate'
      operationId: 7aa034a866ed396d5cb3dc2319ee7ae0
      parameters:
        -
          name: id
          in: path
          description: 'Exchange rate ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Exchange rate deleted successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Exchange rate deleted successfully'
                type: object
        404:
          description: 'Exchange rate not found'
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
...
components:
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...
security:
  -
    sanctum:
      type: apiKey
      description: 'Enter token in format (<token>)'
      name: Authorization
      in: header